% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_hgnc}
\alias{all_hgnc}
\title{all_hgnc}
\usage{
all_hgnc(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
all_hgnc dataset
}
\description{
\code{all_hgnc} returns the all_hgnc dataset
}
\examples{
all_hgnc()
}
