% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alzh_gwas_top100}
\alias{alzh_gwas_top100}
\title{alzh_gwas_top100}
\usage{
alzh_gwas_top100()
}
\value{
alzh_gwas_top100 dataset
}
\description{
\code{alzh_gwas_top100} returns the alzh_gwas_top100 dataset
}
\examples{
alzh_gwas_top100
}
