% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cortex_mrna}
\alias{cortex_mrna}
\title{Karolinska Cortex/Hippocamus dataset subsample}
\format{
An object of class \code{function} of length 1.
}
\source{
\code{
download.file("goo.gl/r5Y24y",destfile="expression_mRNA_17-Aug-2014.txt")
path = "expression_mRNA_17-Aug-2014.txt"
cortex_mrna  = load.linnarsson.sct.data("expression_mRNA_17-Aug-2014.txt")
save(cortex_mrna,file="data/cortex_mrna.rda")
# remove folder once tested
unlink("expression_mRNA_17-Aug-2014.txt")
}
}
\usage{
cortex_mrna
}
\description{
Karolinska Cortex/Hippocamus dataset subsample
}
\examples{
ewceData::cortex_mrna()
}
\keyword{datasets}
