% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hpsd_genes}
\alias{hpsd_genes}
\title{hpsd_genes}
\usage{
hpsd_genes(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
hpsd_genes dataset
}
\description{
\code{hpsd_genes} returns the hpsd_genes dataset
}
\examples{
hpsd_genes()
}
