% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hypothalamus_mrna}
\alias{hypothalamus_mrna}
\title{hypothalamus_mrna}
\usage{
hypothalamus_mrna()
}
\value{
hypothalamus_mrna dataset
}
\description{
\code{hypothalamus_mrna} returns the hypothalamus_mrna dataset
}
\examples{
hypothalamus_mrna()
}
