% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{id_genes}
\alias{id_genes}
\title{Intellectual disability genes}
\format{
An object of class \code{function} of length 1.
}
\source{
from http://compbio.charite.de/hpoweb/showterm?id=HP:0001249
\code{
file3 <- paste0("/Users/ns9/Google Drive/DiseaseLists/",
                     "Intellectual Disability July2017.txt")
id_genes <- read.csv(file3,stringsAsFactors = FALSE)[-1,1]
save(id_genes,file="data/id_genes.rda")
}
}
\usage{
id_genes
}
\description{
Intellectual disability genes
}
\examples{
id_genes <- ewceData::id_genes()
}
\keyword{datasets}
