% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgi_synonym_data}
\alias{mgi_synonym_data}
\title{mgi_synonym_data}
\usage{
mgi_synonym_data(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
mgi_synonym_data dataset
}
\description{
\code{mgi_synonym_data} returns the mgi_synonym_data dataset
}
\examples{
mgi_synonym_data()
}
