% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mouse_to_human_homologs}
\alias{mouse_to_human_homologs}
\title{mouse_to_human_homologs}
\usage{
mouse_to_human_homologs()
}
\value{
mouse_to_human_homologs dataset
}
\description{
\code{mouse_to_human_homologs} returns the mouse_to_human_homologs dataset
}
\examples{
mouse_to_human_homologs()
}
