% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbfox_genes}
\alias{rbfox_genes}
\title{rbfox_genes}
\usage{
rbfox_genes(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
rbfox_genes dataset
}
\description{
\code{rbfox_genes} returns the rbfox_genes dataset
}
\examples{
rbfox_genes()
}
