% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{schiz_genes}
\alias{schiz_genes}
\title{schiz_genes}
\usage{
schiz_genes(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
schiz_genes dataset
}
\description{
\code{schiz_genes} returns the schiz_genes dataset
}
\examples{
schiz_genes()
}
