% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tt_alzh}
\alias{tt_alzh}
\title{tt_alzh}
\usage{
tt_alzh(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
tt_alzh dataset
}
\description{
\code{tt_alzh} returns the tt_alzh dataset
}
\examples{
tt_alzh()
}
