% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tt_alzh_BA44}
\alias{tt_alzh_BA44}
\title{tt_alzh_BA44}
\usage{
tt_alzh_BA44()
}
\value{
tt_alzh_BA44 dataset
}
\description{
\code{tt_alzh_BA44} returns the tt_alzh_BA44 dataset
}
\examples{
tt_alzh_BA44()
}
