% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_hgnc_wtEnsembl}
\alias{all_hgnc_wtEnsembl}
\title{all_hgnc_wtEnsembl}
\usage{
all_hgnc_wtEnsembl()
}
\value{
all_hgnc_wtEnsembl dataset
}
\description{
\code{all_hgnc_wtEnsembl} returns the all_hgnc_wtEnsembl dataset
}
\examples{
all_hgnc_wtEnsembl()
}
