% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_mgi_wtEnsembl}
\alias{all_mgi_wtEnsembl}
\title{all_mgi_wtEnsembl}
\usage{
all_mgi_wtEnsembl()
}
\value{
all_mgi_wtEnsembl dataset
}
\description{
\code{all_mgi_wtEnsembl} returns the all_mgi_wtEnsembl dataset
}
\examples{
all_mgi_wtEnsembl()
}
