% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_mgi_wtEnsembl}
\alias{all_mgi_wtEnsembl}
\title{all_mgi_wtEnsembl}
\usage{
all_mgi_wtEnsembl(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
all_mgi_wtEnsembl dataset
}
\description{
\code{all_mgi_wtEnsembl} returns the all_mgi_wtEnsembl dataset
}
\examples{
all_mgi_wtEnsembl()
}
