% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alzh_gwas_top100}
\alias{alzh_gwas_top100}
\title{alzh_gwas_top100}
\usage{
alzh_gwas_top100(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
alzh_gwas_top100 dataset
}
\description{
\code{alzh_gwas_top100} returns the alzh_gwas_top100 dataset
}
\examples{
alzh_gwas_top100
}
