% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctd}
\alias{ctd}
\title{ctd}
\usage{
ctd(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
ctd dataset
}
\description{
\code{ctd} returns the ctd dataset
}
\examples{
ctd()
}
