% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensembl_transcript_lengths_GCcontent}
\alias{ensembl_transcript_lengths_GCcontent}
\title{ensembl_transcript_lengths_GCcontent}
\usage{
ensembl_transcript_lengths_GCcontent()
}
\value{
ensembl_transcript_lengths_GCcontent dataset
}
\description{
\code{ensembl_transcript_lengths_GCcontent} returns the ensembl_transcript_lengths_GCcontent dataset
}
\examples{
ensembl_transcript_lengths_GCcontent()
}
