% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_genelist}
\alias{example_genelist}
\title{example_genelist}
\usage{
example_genelist()
}
\value{
example_genelist dataset
}
\description{
\code{example_genelist} returns the example_genelist dataset
}
\examples{
example_genelist()
}
