% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_genelist}
\alias{example_genelist}
\title{Example Gene list}
\format{
An object of class \code{function} of length 1.
}
\source{
These were obtained from two sources:
http://www.alzgene.org/TopResults.asp and PMID24162737
}
\usage{
example_genelist
}
\description{
A list of genes genetically associated with Alzheimer's disease.
}
\examples{
example_genelist <- ewceData::example_genelist()
}
\keyword{datasets}
