% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hypothalamus_mrna}
\alias{hypothalamus_mrna}
\title{hypothalamus_mrna}
\usage{
hypothalamus_mrna(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
hypothalamus_mrna dataset
}
\description{
\code{hypothalamus_mrna} returns the hypothalamus_mrna dataset
}
\examples{
hypothalamus_mrna()
}
