% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_genes}
\alias{id_genes}
\title{id_genes}
\usage{
id_genes(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
id_genes dataset
}
\description{
\code{id_genes} returns the id_genes dataset
}
\examples{
id_genes()
}
