% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tt_alzh_BA36}
\alias{tt_alzh_BA36}
\title{tt_alzh_BA36}
\usage{
tt_alzh_BA36()
}
\value{
tt_alzh_BA36 dataset
}
\description{
\code{tt_alzh_BA36} returns the tt_alzh_BA36 dataset
}
\examples{
tt_alzh_BA36()
}
