% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tt_alzh_BA36}
\alias{tt_alzh_BA36}
\title{Example table of differential expression (Alzheimer's BA36)}
\format{
An object of class \code{function} of length 1.
}
\source{
The table was determined based on data associated with the paper with
PMID:17845826
}
\usage{
tt_alzh_BA36
}
\description{
A list of genes found to be differentially expressed in the BA36 in
Alzheimer's disease.
}
\examples{
ewceData::tt_alzh_BA36()
}
\keyword{datasets}
