% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tt_alzh_BA44}
\alias{tt_alzh_BA44}
\title{tt_alzh_BA44}
\usage{
tt_alzh_BA44(localHub = FALSE)
}
\arguments{
\item{localHub}{If working offline, add argument localHub=TRUE to work
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
}
\value{
tt_alzh_BA44 dataset
}
\description{
\code{tt_alzh_BA44} returns the tt_alzh_BA44 dataset
}
\examples{
tt_alzh_BA44()
}
